+++ src/classes/Carts/CartQueue.ts
@@ -1,9 +1,9 @@
 import SteamID from 'steamid';
 
-import Bot from './Bot';
+import Bot from '../Bot';
 import Cart from './Cart';
 
-import log from '../lib/logger';
+import log from '../../lib/logger';
 
 export = CartQueue;
